requirements.txt
streamlit
langchain
openai
python-dotenv
langchain-openai
langchain-community
langchain-openai
